LPS-123793 Refactor the Form component to compose with the new components and reducers
Well FormView is an application of the final result of the form visible to the user. It is also an application that handles reducers differently than Form Builder.

Not all reducers that are used in Forms are used in the User view, so we keep only the reducers that are used and also the proprietary reducers of the User view screen.
