+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-renderer/src/main/java/com/liferay/dynamic/data/mapping/form/renderer/internal/DDMFormRendererImpl.java
@@ -133,7 +133,7 @@ public class DDMFormRendererImpl implements DDMFormRenderer {
 	}
 
 	private static final String _MODULE_NAME =
-		"dynamic-data-mapping-form-renderer/js/containers/Form.es";
+		"dynamic-data-mapping-form-renderer/js/custom/form/FormView.es";
 
 	@Reference
 	private DDM _ddm;
