+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-web/src/main/resources/META-INF/resources/admin/js/index.es.js
@@ -12,6 +12,6 @@
  * details.
  */
 
-import {Form} from 'dynamic-data-mapping-form-renderer';
+import {FormView} from 'dynamic-data-mapping-form-renderer';
 
-export default Form;
+export default FormView;
